Correct nl-nl abbreviations
When speaking about money in dutch we use the full words miljoen, miljard, biljoen even when its written as a abbreviation.
For that reason they need to end with a dot and have a space.
The best abbreviation for miljard is mld.. 
mrd. is possible but their is no matching abbreviation for miljoen (men.), biljoen (ben).