Fix formatting for small numbers.
Fixes #596.  Formatting numbers smaller than approximately 1e-6 resulted
in NaN (instead of a formatted string).  This was due to coercing the
numeric value to a string and adding an exponential notation to that;
however, small numbers coerce to exponential representation anyway, so
it was getting that notation twice.  In other words, 1e-7 would result
in the string 1e-7e+0 (for example), which was not parsing correctly.